+++ front/src/app/features/auth/components/login/login.component.spec.ts
@@ -1,4 +1,4 @@
-import { HttpClient, HttpClientModule, HttpErrorResponse } from '@angular/common/http';
+import { HttpClientModule, HttpErrorResponse } from '@angular/common/http';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { ReactiveFormsModule } from '@angular/forms';
 import { MatCardModule } from '@angular/material/card';
@@ -67,70 +67,6 @@ describe('LoginComponent', () => {
     component.submit();
     fixture.detectChanges();
     expect(component.onError).toBe(true);
-  });
-
-  it('Submit button should be disabled when fields are empty', () => {
-    const submitBtn : HTMLButtonElement | null = fixture.nativeElement.querySelector('button[type="submit"]')
-    expect(submitBtn?.disabled).toBe(true);
-  });
-
-  it('email field should be invalid with wrong email and submit should be disabled', () => {
-    component.form.setValue({email: 'bob-test.com', password: '123456'});
-    fixture.detectChanges();
-    const emailField : HTMLFormElement | null = fixture.nativeElement.querySelector('input[ng-reflect-name="email"]');
-    const submitBtn : HTMLButtonElement | null = fixture.nativeElement.querySelector('button[type="submit"]')
-    expect(submitBtn?.disabled).toBe(true);
-    expect(emailField?.classList).toContain("ng-invalid");
-  });
-
-  it('password field should be invalid when empty and submit should be disabled', () => {
-    component.form.setValue({email: 'bob@test.com', password: ''});
-    fixture.detectChanges();
-    const passField : HTMLFormElement | null = fixture.nativeElement.querySelector('input[ng-reflect-name="password"]');
-    const submitBtn : HTMLButtonElement | null = fixture.nativeElement.querySelector('button[type="submit"]')
-    expect(submitBtn?.disabled).toBe(true);
-    expect(passField?.classList).toContain("ng-invalid");
-  });
-
-  it('Submit button should be enabled when fields are ok', () => {
-    component.form.setValue({email: 'bob@test.com', password: '123456'});
-    fixture.detectChanges();
-    const submitBtn : HTMLButtonElement | null = fixture.nativeElement.querySelector('button[type="submit"]')
-    expect(submitBtn?.disabled).toBe(false);
-  });
-
-  it('click on Submit button with successfull login should login session and go to /session', () => {
-    component.form.setValue({email: 'bob@test.com', password: '123456'});
-    fixture.detectChanges();
-    const submitBtn : HTMLButtonElement | null = fixture.nativeElement.querySelector('button[type="submit"]') 
-    const sessionServiceSpy = jest.spyOn(sessionService,"logIn");
-    const session = Object({
-      token: "token",
-      type: "type",
-      username: "bob",
-      firstName: "Bob",
-      lastName: "Le Bricoleur",
-      admin: true
-    }) as SessionInformation;
-    const routerSpy = jest.spyOn(router,"navigate");
-    authService.login = jest.fn(() => new Observable((obs) => obs.next(session)));
-    submitBtn?.click();
-    expect(sessionServiceSpy).toBeCalledWith(session)
-    expect(routerSpy).toBeCalledWith(['/sessions']);
-    expect(component.onError).toBe(false);
-  });
-
-  it('click on Submit button with login failure should display error', () => {
-    const httpError = new HttpErrorResponse({ error: 'Unhautorized', status: 401 });
-    const loginElement: HTMLElement = fixture.nativeElement;
-    const submitBtn : HTMLButtonElement | null = loginElement.querySelector('button[type="submit"]') 
-    authService.login = jest.fn(() => new Observable((obs) => obs.error(httpError)));
-    component.form.setValue({email: 'bob@test.com', password: '123456'});
-    fixture.detectChanges();
-    expect(submitBtn?.disabled).toBe(false);
-    submitBtn?.click()
-    fixture.detectChanges(); 
-    expect(loginElement.querySelector('p.error')?.textContent).toContain('An error occurred');
   });
 
 });
